API: Make PACKMOL wrappers public
