+++ openff/interchange/_tests/interoperability_tests/components/test_packmol.py
@@ -2,11 +2,11 @@ import pytest
 from openff.toolkit import Quantity
 
 from openff.interchange._tests import MoleculeWithConformer
-from openff.interchange.components._packmol import (
+from openff.interchange.drivers import get_openmm_energies
+from openff.interchange.packmol import (
     _find_packmol,
     solvate_topology_nonwater,
 )
-from openff.interchange.drivers import get_openmm_energies
 
 
 @pytest.mark.slow
